fix preset test
